dns: Refactor and improve Resolver read/receive flow
* Introduce the 'connections' channel to notify new connections.
* Integrate the read() loop into relay(), so whenever a new connection
  created, read() is called to handle it.
* Remove the retry logic from Query() as it's not really necessary.
* Remove the early connect() probe from NewResolver() for two reasons:
  - it's not really necessary; one can even argue it's better to do
    nothing unless necessary.
  - it requires to create the 'connections' channel with buffer (i.e.,
    given a size).
* Remove the 'responses' channel by directly passing the 'forwarder'
  channel to read().

In addition, this refactoring also fixed this bug:
When the very first connection created in NewResolver() timeout (e.g.,
no query happens in the first several seconds), then the first query
would fail because the read() fails due to the connection closed by the
remote.  In the other hand, a later query would trigger the reconnection
and thus succeed.  The refactored read() flow no longer has such a bug.

However, note that this single shared connection method is actually
better suited for UDP forwarding, while for this TCP/DoT forwarding,
it's better to use a TCP connection pool with keepalive.  Will work on
this later.